Refactor

+++ src/tokenizer.cpp
@@ -217,7 +217,7 @@ void Tokenizer::TokenStream::addToken(TokenKind kind, char* pos){
 }
 
 bool Tokenizer::TokenStream::consume(TokenKind kind){
-    if(idx >= tokens.size()){
+    if(idx >= (int32_t)tokens.size()){
         return false;
     }
     if(tokens[idx].kind != kind){
@@ -229,7 +229,7 @@ bool Tokenizer::TokenStream::consume(TokenKind kind){
 }
 
 void Tokenizer::TokenStream::expect(TokenKind kind){
-    if(idx >= tokens.size()){
+    if(idx >= (int32_t)tokens.size()){
         fprintf(stderr, "Unexpected end of input\n");
         exit(1);
     }
@@ -244,7 +244,7 @@ void Tokenizer::TokenStream::expect(TokenKind kind){
 }
 
 int32_t Tokenizer::TokenStream::expectNum(){
-    if(idx >= tokens.size()){
+    if(idx >= (int32_t)tokens.size()){
         fprintf(stderr, "Unexpected end of input\n");
         exit(1);
     }
@@ -258,7 +258,7 @@ int32_t Tokenizer::TokenStream::expectNum(){
 }
 
 TokenIdx Tokenizer::TokenStream::expectIdent(){
-    if(idx >= tokens.size()){
+    if(idx >= (int32_t)tokens.size()){
         fprintf(stderr, "Unexpected end of input\n");
         exit(1);
     }
@@ -272,7 +272,7 @@ TokenIdx Tokenizer::TokenStream::expectIdent(){
 }
 
 TokenIdx Tokenizer::TokenStream::expectStringLiteral(){
-    if(idx >= tokens.size()){
+    if(idx >= (int32_t)tokens.size()){
         fprintf(stderr, "Unexpected end of input\n");
         exit(1);
     }
@@ -286,7 +286,7 @@ TokenIdx Tokenizer::TokenStream::expectStringLiteral(){
 }
 
 std::optional<TokenIdx> Tokenizer::TokenStream::consumeIdent(){
-    if(idx >= tokens.size()){
+    if(idx >= (int32_t)tokens.size()){
         return std::nullopt;
     }
     if(tokens[idx].kind != TokenKind::Ident){
